+++ server/src/main/java/com/StoreX/common/datatypes/bo/PozycjaWydaniaBO.java
@@ -1,19 +1,16 @@
 package com.StoreX.common.datatypes.bo;
 
-import com.StoreX.persistence.entity.Towar;
-import com.StoreX.persistence.entity.WydanieZamowienia;
-
 public class PozycjaWydaniaBO {
-    private int ID;
+    private Long ID;
     private double ilosc;
     private TowarBO towar;
     private WydanieZamowieniaBO wydanieZamowienia;
 
-    public int getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
