+++ server/src/main/java/com/StoreX/common/datatypes/bo/WydanieZamowieniaBO.java
@@ -1,18 +1,16 @@
 package com.StoreX.common.datatypes.bo;
-import com.StoreX.persistence.entity.ZamowienieZakupu;
-
 import java.util.Date;
 
 public class WydanieZamowieniaBO {
-    private int ID;
+    private Long ID;
     private Date data;
     private ZamowienieZakupuBO zamowienie;
 
-    public int getID() {
+    public Long getID() {
         return ID;
     }
 
-    public void setID(int ID) {
+    public void setID(Long ID) {
         this.ID = ID;
     }
